Move BlueZGattApplication to dbus module
