+++ bless/backends/bluezdbus/advertisement.py
@@ -6,7 +6,7 @@ from txdbus.objects import DBusObject, DBusProperty, dbusMethod  # type: ignore
 from txdbus.interface import DBusInterface, Method, Property  # type: ignore
 
 if TYPE_CHECKING:
-    from bless.backends.bluezdbus.application import (  # type: ignore
+    from bless.backends.bluezdbus.dbus.application import (  # type: ignore
         BlueZGattApplication,
     )
