+++ bless/backends/bluezdbus/server.py
@@ -15,8 +15,10 @@ from bleak.backends.bluezdbus.characteristic import (  # type: ignore
 )
 
 from bless.backends.server import BaseBlessServer  # type: ignore
-from bless.backends.bluezdbus.application import BlueZGattApplication  # type: ignore
 from bless.backends.bluezdbus.characteristic import flags_to_dbus
+from bless.backends.bluezdbus.dbus.application import (  # type: ignore
+        BlueZGattApplication
+        )
 from bless.backends.bluezdbus.dbus.service import BlueZGattService  # type: ignore
 from bless.backends.bluezdbus.dbus.utils import get_adapter  # type: ignore
 from bless.backends.bluezdbus.dbus.characteristic import (  # type: ignore
@@ -25,9 +27,9 @@ from bless.backends.bluezdbus.dbus.characteristic import (  # type: ignore
 )
 
 from bless.backends.characteristic import (  # type: ignore
-        GATTCharacteristicProperties,
-        GATTAttributePermissions
-        )
+    GATTCharacteristicProperties,
+    GATTAttributePermissions,
+)
 
 
 class BlessServerBlueZDBus(BaseBlessServer):
