+++ test/backends/bluezdbus/test_application.py
@@ -16,7 +16,7 @@ from txdbus.objects import RemoteDBusObject  # type: ignore # noqa: E402
 
 from bless.backends.bluezdbus.dbus.characteristic import Flags  # type: ignore # noqa: E402 E501
 from bless.backends.bluezdbus.dbus.utils import get_adapter  # type: ignore # noqa: E402 E501
-from bless.backends.bluezdbus.application import BlueZGattApplication  # type: ignore # noqa: E402 E501
+from bless.backends.bluezdbus.dbus.application import BlueZGattApplication  # type: ignore # noqa: E402 E501
 from bless.backends.bluezdbus.dbus.characteristic import BlueZGattCharacteristic  # type: ignore # noqa: E402 E501
 
 from twisted.internet.asyncioreactor import AsyncioSelectorReactor  # type: ignore # noqa: E402 E501
